requirements.txt
requests==2.32.4
selenium==4.15.1
reportlab==3.6.13
python-dotenv==1.0.0
pypdf==3.17.0